Completed "Physics - Part 2" of the course.
I have created the crate object as part of this lesson, which is just an object with a rigidbody. One small difference from the lesson that I did make is that I didn't zoom in the camera and I added dampening to angular velocity so that my crates don't spin ad infinitum.